requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cvxpy==1.7.4
cvxpylayers==0.1.9
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
diffcp==1.1.8
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.10.0
gekko==1.3.0
gitdb==4.0.12
GitPython==3.1.45
gurobipy==12.0.0
hydra-core==1.3.2
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
osqp==1.0.5
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.7.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
PYPOWER==5.1.19
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.7.2
scikit-optimize==0.10.2
scipy==1.15.3
scs==3.2.9
sentry-sdk==2.43.0
setproctitle==1.3.7
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.22.3
wcwidth==0.2.14
